requirements.txt
# Consult release notes for supported versions of Django, Wagtail, and Python.
# https://docs.coderedcorp.com/wagtail-crx/releases/
coderedcms==4.0.*
wagtail==6.3.*
django==5.0.9
setuptools
fiona
pyproj
geos
gdal
shapely
djangorestframework
djangorestframework-gis
django-auditlog
django-autoslug
django-sass
django-json-widget
django-guardian
django-allauth
django-polymorphic
django-modelcluster
django-admin-rangefilter
django-filter
django-cors-headers
wagtail-rangefilter
djoser
earthengine-api
httpx
python-dotenv

# pygeoapi
pygeoapi
pydantic<2.0.0
jinja2
Babel
click
filelock
jsonschema
pygeofilter
pygeoif
python-dateutil
pytz
PyYAML
rasterio
requests
SQLAlchemy<2.0.0
tinydb
unicodecsv
jsonpatch

elasticsearch
elasticsearch_dsl